moved components in prep for ui improvements
